Fixes FF daily data with different data/exchange tz
The issue with having a different data and exchange time zone is that
daily data no longer follows a common assumption of beginning and ending
at midnight each day. This assumption was baked into the existing impl
to its own detriment. Specifically, the get next market open date calc
was being improperly executed. In most case (data resolution < one day)
we would use the start time of the bar, but since daily data spans a day,
we need to use the end time of the bar to properly calculate this value.

Fixes #1073